backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.25
asyncpg==0.29.0

# --- RAG & LLM (Pinned Stable Versions) ---
langchain==0.3.0
langchain-community==0.3.0
langchain-core==0.3.0
langchain-text-splitters==0.3.0
langchain-ollama==0.2.0
langchain-huggingface==0.1.0

sentence-transformers==2.7.0
faiss-cpu==1.8.0
python-dotenv
tzdata

langchain-groq>=0.1.0